+++ hack/update-generated-deep-copies.sh
@@ -16,14 +16,14 @@ ALL_FQ_APIS=(
     github.com/openshift/origin/pkg/cmd/server/apis/config/v1/testing
     github.com/openshift/origin/pkg/image/apiserver/admission/apis/imagepolicy
     github.com/openshift/origin/pkg/image/apiserver/admission/apis/imagepolicy/v1
-    github.com/openshift/origin/pkg/network/apiserver/admission/apis/ingressadmission
-    github.com/openshift/origin/pkg/network/apiserver/admission/apis/ingressadmission/v1
     github.com/openshift/origin/pkg/project/apiserver/admission/apis/requestlimit
     github.com/openshift/origin/pkg/project/apiserver/admission/apis/requestlimit/v1
     github.com/openshift/origin/pkg/quota/apiserver/admission/apis/clusterresourceoverride
     github.com/openshift/origin/pkg/quota/apiserver/admission/apis/clusterresourceoverride/v1
     github.com/openshift/origin/pkg/quota/apiserver/admission/apis/runonceduration
     github.com/openshift/origin/pkg/quota/apiserver/admission/apis/runonceduration/v1
+    github.com/openshift/origin/pkg/route/apiserver/admission/apis/ingressadmission
+    github.com/openshift/origin/pkg/route/apiserver/admission/apis/ingressadmission/v1
     github.com/openshift/origin/pkg/router/f5/testing
     github.com/openshift/origin/pkg/scheduler/admission/apis/podnodeconstraints
     github.com/openshift/origin/pkg/scheduler/admission/apis/podnodeconstraints/v1
