+++ pkg/cmd/server/apis/config/install/install.go
@@ -11,10 +11,10 @@ import (
 	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
 	configapiv1 "github.com/openshift/origin/pkg/cmd/server/apis/config/v1"
 	imagepolicyinstall "github.com/openshift/origin/pkg/image/apiserver/admission/apis/imagepolicy/install"
-	ingressadmissioninstall "github.com/openshift/origin/pkg/network/apiserver/admission/apis/ingressadmission/install"
 	requestlimitinstall "github.com/openshift/origin/pkg/project/apiserver/admission/apis/requestlimit/install"
 	clusterresourceoverrideinstall "github.com/openshift/origin/pkg/quota/apiserver/admission/apis/clusterresourceoverride/install"
 	runoncedurationinstall "github.com/openshift/origin/pkg/quota/apiserver/admission/apis/runonceduration/install"
+	ingressadmissioninstall "github.com/openshift/origin/pkg/route/apiserver/admission/apis/ingressadmission/install"
 	podnodeconstraintsinstall "github.com/openshift/origin/pkg/scheduler/admission/apis/podnodeconstraints/install"
 	externaliprangerinstall "github.com/openshift/origin/pkg/service/admission/apis/externalipranger/install"
 	restrictedendpointsinstall "github.com/openshift/origin/pkg/service/admission/apis/restrictedendpoints/install"
