+++ pkg/route/apiserver/admission/apis/ingressadmission/install/install.go
@@ -4,8 +4,8 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 
-	"github.com/openshift/origin/pkg/network/apiserver/admission/apis/ingressadmission"
-	"github.com/openshift/origin/pkg/network/apiserver/admission/apis/ingressadmission/v1"
+	"github.com/openshift/origin/pkg/route/apiserver/admission/apis/ingressadmission"
+	"github.com/openshift/origin/pkg/route/apiserver/admission/apis/ingressadmission/v1"
 )
 
 func InstallLegacyInternal(scheme *runtime.Scheme) {
